server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.1
aiofiles==23.2.1
llama-cpp-python==0.2.26
sentencepiece==0.2.0
protobuf==3.20.3

# For model loading - choose one based on your model type
# For GGUF models (recommended for CPU):
llama-cpp-python==0.2.26

# For Hugging Face transformers models:
# transformers==4.35.0
# torch==2.1.0
# accelerate==0.24.1